docker/kafka/kafka-benchmark/requirements.txt
kafka-python
beautifulsoup4
googlesearch-python
nltk